requirements.txt
django
pillow
pandas
numpy
cv2
psycopg2
xlwt
xlutils
opencv-python
opencv-contrib-python
